requirements/base.txt
# web
Django==1.10.1
djangorestframework==3.4.6
django-selectable==0.9.0
django-extensions==1.7.3
django-webpack-loader==0.3.3
django-redis
django-picklefield==1.0.0
django-crispy-forms==1.7.0

# tasks
celery==4.1.0
kombu==4.1.0

# utilities
click==6.6
psycopg2==2.6.2
redis==2.10.5
fuzzywuzzy==0.15.1
requests==2.12.0
parse==1.8.2
progress==1.3
mysql-connector==2.1.4

# computational
fa2==0.2
gensim==2.3.0
h5py==2.8.0
Keras==2.1.4
matplotlib==1.5.3
networkx==2.1
nltk==3.2.4
numpy==1.12.1
pandas==0.19.2
pyBigWig==0.2.8
python-Levenshtein==0.12.0
scipy==0.19.0
scikit-learn==0.18.1
sklearn==0.0
tensorflow==1.6.0
